+++ IHS300_AnomalyDetection/read_csv.h
@@ -14,12 +14,12 @@
 #include <string.h>
 
 // Read constant value from csv file
-double load_csv_constant(const char* filename);
+float load_csv_constant(const char* filename);
 
 // Read 1D array from csv file
-void load_csv_1d(const char* filename, double** data, int* size);
+void load_csv_1d(const char* filename, float** data, int* size);
 
 // Read 2D array from csv file
-void load_csv_2d(const char* filename, double*** data, int* rows, int* cols);
+void load_csv_2d(const char* filename, float*** data, int* rows, int* cols);
 
 #endif
